% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_build.R
\name{dev_build_readme}
\alias{dev_build_readme}
\title{Build README from README.Rmd}
\usage{
dev_build_readme(pkg = ".", quiet = FALSE)
}
\arguments{
\item{pkg}{Path to package directory. Default is current directory.}

\item{quiet}{Logical. If TRUE, suppress console output. Default FALSE.}
}
\value{
Path to README.md invisibly.
}
\description{
Renders README.Rmd to README.md by executing all R code chunks and
updating output. Useful for keeping examples and badges up-to-date.
}
\examples{
\dontrun{
# Build README
dev_build_readme()
}
}
