% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blankpkg_dummy.R
\name{blank_x}
\alias{blank_x}
\title{Dummy placeholder function}
\usage{
blank_x(input = "planceholder code")
}
\arguments{
\item{input}{character string}
}
\value{
echo the input string
}
\description{
This is a small dummy function to exemplify the
{roxygen} documentation, the automatic testing etc...
}
\examples{
blank_x()

blank_x("something else...")
}
